Make container random question everytime on pushing new screen
